Compund glyphs + cmap
Compund glyph reading supported (apart from anchor point positioning) as well as character mapping (formats 4, 12 and 13)

+++ TextRenderer/Assets/Code/FileReadUtilities.cs
@@ -3,7 +3,7 @@ using System;
 
 public static class FileReadUtilities
 {
-    public static void SkipBytes(this Stream stream, int count)
+    public static void SkipBytes(this Stream stream, long count)
     {
         stream.Seek(count, SeekOrigin.Current);
     }
